+++ lib/imgproxy-rails/helpers.rb
@@ -2,9 +2,13 @@
 
 module ImgproxyRails
   module Helpers
-    def self.image_variation?(model)
-      model.respond_to?(:variation) &&
-        model.try(:blob)&.content_type&.split("/")&.first == "image"
+    def self.applicable_variation?(model)
+      return false if !model.respond_to?(:variation)
+
+      content_type = model.try(:blob)&.content_type
+      content_type&.start_with?("image/") ||
+        content_type&.start_with?("video/") ||
+        content_type == "application/pdf"
     end
   end
 end